requirements.txt
pytest==8.0.2
playwright==1.42.0
pytest-playwright==0.5.2
allure-pytest==2.13.3
requests==2.31
